+++ src/built-in-plugins/command-dev/dashboard/client/stores/application.ts
@@ -1,14 +1,32 @@
-import { message } from "antd"
+import { Icon, message } from "antd"
 import { Action, inject, observable } from "dob"
 import * as React from "react"
 import * as io from "socket.io-client"
 import { IProjectStatus } from "../../server/project-status-interface"
+import { Event } from "../utils/event"
+
+const TreeIcon = (props: any) =>
+  React.createElement(Icon, {
+    ...props,
+    style: {
+      marginRight: 5
+    }
+  })
 
 const serverPort = (window as any)["serverPort"]
+
+interface ITreeNode {
+  children?: ITreeNode[]
+  key: string
+  title: string
+  icon?: React.ReactElement<any>
+  disabled?: boolean
+}
 
 interface IPlugin {
   position: string
   view: any
+  init?: (applicationAction: ApplicationAction) => void
 }
 
 @observable
@@ -25,9 +43,15 @@ export class ApplciationStore {
    * Plugins
    */
   public plugins: IPlugin[] = []
+  /**
+   * Project tree data
+   */
+  public treeData: ITreeNode[] = []
 }
 
 export class ApplicationAction {
+  public event = new Event()
+
   @inject(ApplciationStore) public applicationStore: ApplciationStore
 
   private socket = io(`//localhost:${serverPort}`)
@@ -37,6 +61,8 @@ export class ApplicationAction {
     this.socket.on("freshProjectStatus", (data: IProjectStatus) => {
       Action(() => {
         this.applicationStore.status = data
+        this.clearTreeData()
+        this.event.emit("freshProjectStatus")
       })
     })
 
@@ -90,10 +116,22 @@ export class ApplicationAction {
   }
 
   @Action
+  public pipeTreeNode(callback: (treeData: ITreeNode[]) => ITreeNode[]) {
+    this.applicationStore.treeData = callback(this.applicationStore.treeData)
+  }
+
+  @Action
   public loadUiPlugins(plugins: IPlugin[]) {
     if (!plugins) {
       return
     }
+
+    // Run plugins init function.
+    plugins.forEach(plugin => {
+      if (plugin.init) {
+        plugin.init(this)
+      }
+    })
 
     plugins.forEach(plugin => this.applicationStore.plugins.push(plugin))
   }
@@ -103,5 +141,19 @@ export class ApplicationAction {
     return this.applicationStore.plugins.filter(plugin => plugin.position === position).map((plugin, index) => {
       return React.createElement(plugin.view, { key: index, ...props })
     })
+  }
+
+  @Action
+  private clearTreeData() {
+    this.applicationStore.treeData = [
+      {
+        title: "Project",
+        key: "project-root",
+        icon: React.createElement(TreeIcon, {
+          type: "chrome"
+        }),
+        children: []
+      }
+    ]
   }
 }