+++ src/node/dev-service.ts
@@ -1,6 +1,6 @@
 import { plugin } from "../utils/plugins"
 
-export const on = (name: string, callback: any) => {
+export const on = (name: string, callback: (data?: any, resolve?: any, reject?: any) => void) => {
   plugin.devServices.socketListeners.push({
     name,
     callback