+++ changelog.md
@@ -2,4 +2,4 @@
 
 ## 0.5.0
 
-* **Break change.** Remove built-in data manager [dob](https://github.com/dobjs/dob), but provide it as a plugin: [pri-plugin-dob](https://github.com/prijs/pri-plugin-dob)
+* **Break change.** Remove built-in data manager [dob](https://github.com/dobjs/dob), but provide it as a plugin: [pri-plugin-dob](https://github.com/prijs/plugins/pri-plugin-dob)